+++ replay-control-app/src/main.rs
@@ -531,6 +531,7 @@ mod ssr {
         server_fn::axum::register_explicit::<replay_control_app::server_fns::GetAllGenres>();
         server_fn::axum::register_explicit::<replay_control_app::server_fns::GetSystemGenres>();
         server_fn::axum::register_explicit::<replay_control_app::server_fns::RandomGame>();
+        server_fn::axum::register_explicit::<replay_control_app::server_fns::RandomGameForSystem>();
         server_fn::axum::register_explicit::<replay_control_app::server_fns::SearchByDeveloper>();
         server_fn::axum::register_explicit::<replay_control_app::server_fns::GetDeveloperGames>();
         server_fn::axum::register_explicit::<replay_control_app::server_fns::GetDeveloperGenres>();