+++ replay-control-app/tests/common/mod.rs
@@ -154,6 +154,7 @@ pub fn register_server_fns() {
     server_fn::axum::register_explicit::<server_fns::GetAllGenres>();
     server_fn::axum::register_explicit::<server_fns::GetSystemGenres>();
     server_fn::axum::register_explicit::<server_fns::RandomGame>();
+    server_fn::axum::register_explicit::<server_fns::RandomGameForSystem>();
     server_fn::axum::register_explicit::<server_fns::GetRegionPreference>();
     server_fn::axum::register_explicit::<server_fns::GetRecommendations>();
     server_fn::axum::register_explicit::<server_fns::RebuildCorruptLibrary>();